Does not overwrite index if the environment variable is set, added excludes to unsupported nodes, and removed some defaults that weren't being used

+++ start.sh
@@ -5,7 +5,7 @@ then
 	#cp /tmp/public/* /usr/src/workspace/public
 fi
 
-if [ ! -f /usr/src/workspace/public/index.html ]
+if [ ! -f /usr/src/workspace/public/index.html ] && [ ! -e "NO_INDEX" ]
 	then
 		cp /tmp/public/index.html /usr/src/workspace/public/index.html
 fi
